Inserted a counter that make a deadline miss every 10 seconds
As talked with the professor we have took away sigprocmask. What is happening in the while loop is the following:
 - If the counter is not divisible for 10 then we go to the next command
 - Sigwait block the execution of the process until a SIGALRM is deployed(every period of clock)
 - When sigwait receive a SIGALMR make the execution continue and therefor the funciton print a text (new period)
 - In case the counter is divisile for 2 then we sleep for 2  seconds, so we don t call sigwait, in those 10 second a SIGALMR is deployed by the clock.
Therefor the newly defined handler is called and a "Missed deadline" message is displayed. The fact that a message is received and hanlded make the programm skip the reamining part of the sleep command (1 sec).
- The next command is sigwait so the function wait another signal from the clock